chore: update bun package manager version and remove bun.lockb
fix: add logger dependency to agent-engine package

feat: enhance Kimi model retrieval with type inference

refactor: remove pipeline-engine dependency from db-schema

fix: update imports in pipelines_table to use schemas

chore: update typescript-config dependency to workspace references in multiple packages

feat: migrate schemas to new structure in packages/schemas

feat: add pipeline-related schemas including BuiltinNodeTypeSchema, CodeFileNodeDataSchema, and others

chore: update package.json files across various packages to use workspace references for typescript-config
