+++ package.json
@@ -43,5 +43,5 @@
   "engines": {
     "node": ">=20"
   },
-  "packageManager": "bun@1.0.21"
+  "packageManager": "bun@1.3.11"
 }
