+++ packages/logger/package.json
@@ -17,7 +17,7 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@repo/typescript-config": "*",
+    "@repo/typescript-config": "workspace:*",
     "pino-pretty": "^13.1.3",
     "typescript": "5.9.2"
   }
