+++ packages/obs/package.json
@@ -12,9 +12,12 @@
     "lint": "oxlint --config ./.oxlintrc.json src/",
     "quality": "bun run check-types && bun run lint"
   },
-  "dependencies": {},
+  "dependencies": {
+    "@repo/db-schema": "workspace:*",
+    "@repo/models": "workspace:*"
+  },
   "devDependencies": {
-    "@repo/typescript-config": "*",
+    "@repo/typescript-config": "workspace:*",
     "@types/node": "^25.6.0",
     "typescript": "5.9.2"
   }
