+++ packages/pipeline-engine/package.json
@@ -17,12 +17,14 @@
   "dependencies": {
     "@repo/obs": "workspace:*",
     "@repo/plugin": "workspace:*",
+    "@repo/schemas": "workspace:*",
+    "@repo/utils": "workspace:*",
     "@types/node": "^25.6.0",
     "neverthrow": "^8.2.0",
     "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@repo/typescript-config": "*",
+    "@repo/typescript-config": "workspace:*",
     "typescript": "5.9.2",
     "vitest": "^4.1.4"
   }
