+++ packages/pipeline-engine/src/schemas/FolderNodeDataSchema.ts
@@ -1,13 +1,1 @@
-import { z } from "zod/v4";
-import { DisclosureModeSchema } from "./DisclosureModeSchema";
-
-export const FolderNodeDataSchema = z.object({
-  label: z.string(),
-  nodeType: z.literal("folder"),
-  folderPath: z.string(),
-  excludedPaths: z.array(z.string()).optional(),
-  disclosureMode: DisclosureModeSchema.optional(),
-  includedExtensions: z.array(z.string()).optional(),
-  description: z.string().optional(),
-});
-export type FolderNodeData = z.infer<typeof FolderNodeDataSchema>;
+export { FolderNodeDataSchema, type FolderNodeData } from "@repo/schemas";
