+++ packages/pipeline-engine/src/schemas/NodeRunStatusSchema.ts
@@ -1,11 +1,1 @@
-import { z } from "zod/v4";
-
-export const NODE_RUN_STATUS_ENUM = {
-  IDLE: "idle",
-  RUNNING: "running",
-  PASS: "pass",
-  FAIL: "fail",
-} as const;
-
-export const NodeRunStatusSchema = z.enum(NODE_RUN_STATUS_ENUM);
-export type NodeRunStatus = z.infer<typeof NodeRunStatusSchema>;
+export { NODE_RUN_STATUS_ENUM, NodeRunStatusSchema, type NodeRunStatus } from "@repo/schemas";
