+++ packages/pipeline-engine/src/schemas/OutputProjectPathNodeDataSchema.ts
@@ -1,10 +1,1 @@
-import { z } from "zod/v4";
-
-export const OutputProjectPathNodeDataSchema = z.object({
-  label: z.string(),
-  nodeType: z.literal("output-project-path"),
-  projectId: z.string().optional(),
-  path: z.string(),
-  description: z.string().optional(),
-});
-export type OutputProjectPathNodeData = z.infer<typeof OutputProjectPathNodeDataSchema>;
+export { OutputProjectPathNodeDataSchema, type OutputProjectPathNodeData } from "@repo/schemas";
