+++ packages/pipeline-engine/src/schemas/PipelineEdgeDataSchema.ts
@@ -1,7 +1,1 @@
-import { z } from "zod/v4";
-
-export const PipelineEdgeDataSchema = z.object({
-  label: z.string().optional(),
-  dataType: z.string().optional(),
-});
-export type PipelineEdgeData = z.infer<typeof PipelineEdgeDataSchema>;
+export { PipelineEdgeDataSchema, type PipelineEdgeData } from "@repo/schemas";
