+++ packages/plugins/package.json
@@ -18,7 +18,7 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@repo/typescript-config": "*",
+    "@repo/typescript-config": "workspace:*",
     "@types/node": "^25.6.0",
     "typescript": "5.9.2",
     "vitest": "^4.1.4"
