+++ packages/schemas/src/pipeline/PipelineEdgeDataSchema.ts
@@ -1,0 +1,7 @@
+import { z } from "zod/v4";
+
+export const PipelineEdgeDataSchema = z.object({
+  label: z.string().optional(),
+  dataType: z.string().optional(),
+});
+export type PipelineEdgeData = z.infer<typeof PipelineEdgeDataSchema>;
