+++ packages/schemas/src/pipeline/PipelineSchema.ts
@@ -1,0 +1,17 @@
+import { z } from "zod/v4";
+import { PipelineNodeSchema } from "./PipelineNodeSchema";
+import { PipelineEdgeSchema } from "./PipelineEdgeSchema";
+
+export const PipelineSchema = z.object({
+  id: z.string(),
+  name: z.string(),
+  description: z.string().default(""),
+  tags: z.array(z.string()),
+  timeoutMs: z.number().nullable(),
+  createdAt: z.coerce.date(),
+  updatedAt: z.coerce.date(),
+  nodes: z.array(PipelineNodeSchema),
+  edges: z.array(PipelineEdgeSchema),
+});
+
+export type PipelineData = z.infer<typeof PipelineSchema>;
