+++ packages/services/package.json
@@ -30,7 +30,7 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
-    "@repo/typescript-config": "*",
+    "@repo/typescript-config": "workspace:*",
     "@types/jszip": "^3.4.1",
     "typescript": "5.9.2",
     "vitest": "^4.1.4"
