+++ packages/ui/package.json
@@ -42,7 +42,7 @@
     "tailwind-merge": "^3.1.0"
   },
   "devDependencies": {
-    "@repo/typescript-config": "*",
+    "@repo/typescript-config": "workspace:*",
     "@types/node": "^22.15.3",
     "@types/react": "19.2.2",
     "@types/react-dom": "19.2.2",
