+++ packages/utils/package.json
@@ -17,7 +17,7 @@
     "neverthrow": "^8.2.0"
   },
   "devDependencies": {
-    "@repo/typescript-config": "*",
+    "@repo/typescript-config": "workspace:*",
     "@types/node": "^25.6.0",
     "typescript": "5.9.2",
     "vitest": "^4.1.4"
